Mock ip_addrs() in utils/minions.py unit test (#443)
Previously the test used `salt.utils.network.ip_addrs()' in the same way
that the tested code did. This worked well as long as at least one IP
address was returned by `salt.utils.network.ip_addrs()'.

Since this is a unit test, it should not depend on the environment,
it should just work™, even if there are no real IP addresses assigned to
the system (or container) that runs the test.

Co-authored-by: Pablo Suárez Hernández <psuarezhernandez@suse.com>