Skip package names without colon (bsc#1208691) (#578)
Fixes a problem in `_find_ptf_packages()` when passing multiple packages to `zypperpkg.remove` / `zypperpkg.purge`. The problem occurs when a passed package is not installed, in that case the output of the `rpm` subprocess is not parsed correctly.

BACKPORT-UPSTREAM=https://github.com/saltstack/salt/pull/63460
DOWNSTREAM-REF=https://github.com/openSUSE/salt/pull/578