Avoid conflicts with dependencies versions (bsc#1211612) (#581)
This commit fixes the Salt requirements file that are used to
generate the "requires.txt" file that is included in Salt egginfo
in order to be consistent with the installed packages
of Salt dependencies.

This prevents issues when resolving and validating Salt dependencies
with "pkg_resources" Python module.

DOWNSTREAM-REF=https://github.com/openSUSE/salt/pull/581